LayoutTests/fast/css/CSSStyleDeclaration-cssText-null.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that CSSStyleDeclaration.cssText is not nullable");

var div = document.createElement("div");
var style = div.style;
shouldBeEqualToString("style.cssText", "");
shouldNotThrow("style.cssText = 'margin: 0px;'");
shouldBeEqualToString("style.cssText", "margin: 0px;");
shouldNotThrow("style.cssText = null");
shouldBeEqualToString("style.cssText", "");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
