Safari-609.1.11.1/LayoutTests/fast/css/CSSStyleDeclaration-cssText-null-expected.txt

Tests that CSSStyleDeclaration.cssText is not nullable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.cssText is ""
PASS style.cssText = 'margin: 0px;' did not throw exception.
PASS style.cssText is "margin: 0px;"
PASS style.cssText = null did not throw exception.
PASS style.cssText is ""
PASS successfullyParsed is true

TEST COMPLETE
